refactored and moved user state upstream
